requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
lxml==4.6.4
pymongo==3.12.1
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
tinydb==4.5.2
urllib3==1.26.7
